@@ -1,12 +1,12 @@
 {
   "name": "atlas-api",
-  "version": "0.0.2",
+  "version": "0.0.3",
   "description": "Atlas RESTful API",
   "author": "Yassire Elhani <yassire@makeandship.com>",
   "main": "src/index.js",
   "private": true,
   "engines": {
-    "node": "8.12.0",
+    "node": "12.13.0",
     "npm": ">=6",
     "yarn": ">=1"
   },
