@@ -9,34 +9,41 @@ import Experiment from "../../src/server/models/experiment.model";
 
 import DataHelper from "../../src/server/helpers/DataHelper";
 
-const app = createApp();
-const users = require("../fixtures/users");
-const experiments = require("../fixtures/experiments");
-let token = null;
+import users from "../fixtures/users";
+import experiments from "../fixtures/experiments";
+
+const args = {
+  app: null,
+  token: null
+};
+
+beforeAll(async () => {
+  args.app = await createApp();
+});
 
 beforeEach(async done => {
   const userData = new User(users.admin);
 
   const savedUser = await userData.save();
-  request(app)
+  request(args.app)
     .post("/auth/login")
     .send({ username: "admin@nhs.co.uk", password: "password" })
     .end(async (err, res) => {
-      token = res.body.data.access_token;
+      args.token = res.body.data.access_token;
       done();
     });
 });
 
 afterEach(async done => {
-  await User.remove({});
-  await Experiment.remove({});
+  await User.deleteMany({});
+  await Experiment.deleteMany({});
   done();
 });
 
 describe("DataController", () => {
   describe("POST /data/create", () => {
     it("should return a successful response", done => {
-      request(app)
+      request(args.app)
         .post("/data/create")
         .send({ total: 1 })
         .expect(httpStatus.OK)
@@ -47,7 +54,7 @@ describe("DataController", () => {
         });
     });
     it("should create 5 users", done => {
-      request(app)
+      request(args.app)
         .post("/data/create")
         .send({ total: 1 })
         .expect(httpStatus.OK)
@@ -60,7 +67,7 @@ describe("DataController", () => {
     });
     it("should create the given number of experiments", done => {
       const random = faker.random.number({ min: 1, max: 5 });
-      request(app)
+      request(args.app)
         .post("/data/create")
         .send({ total: random })
         .expect(httpStatus.OK)
@@ -72,7 +79,7 @@ describe("DataController", () => {
         });
     });
     it("should populate the experiment file", done => {
-      request(app)
+      request(args.app)
         .post("/data/create")
         .send({ total: 1 })
         .expect(httpStatus.OK)
@@ -83,7 +90,7 @@ describe("DataController", () => {
         });
     });
     it("should populate the experiment metadata", done => {
-      request(app)
+      request(args.app)
         .post("/data/create")
         .send({ total: 1 })
         .expect(httpStatus.OK)
@@ -94,7 +101,7 @@ describe("DataController", () => {
         });
     });
     it("should populate experiment results", done => {
-      request(app)
+      request(args.app)
         .post("/data/create")
         .send({ total: 1 })
         .expect(httpStatus.OK)
@@ -124,7 +131,7 @@ describe("DataController", () => {
         });
     });
     it("should populate the experiment owner", done => {
-      request(app)
+      request(args.app)
         .post("/data/create")
         .send({ total: 1 })
         .expect(httpStatus.OK)
@@ -138,7 +145,7 @@ describe("DataController", () => {
 
   describe("POST /data/clean", () => {
     it("should return a successful response", done => {
-      request(app)
+      request(args.app)
         .post("/data/clean")
         .expect(httpStatus.OK)
         .end((err, res) => {
@@ -148,7 +155,7 @@ describe("DataController", () => {
         });
     });
     it("should clear all users", done => {
-      request(app)
+      request(args.app)
         .post("/data/clean")
         .expect(httpStatus.OK)
         .end(async (err, res) => {
@@ -159,7 +166,7 @@ describe("DataController", () => {
         });
     });
     it("should clear all experiments", done => {
-      request(app)
+      request(args.app)
         .post("/data/clean")
         .expect(httpStatus.OK)
         .end(async (err, res) => {
@@ -171,202 +178,196 @@ describe("DataController", () => {
     });
   });
 
-  describe("POST /data/demo/:folder", () => {
-    describe("when passing invalid data", () => {
-      it("should be a protected route", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", "Bearer INVALID")
-          .expect(httpStatus.UNAUTHORIZED)
-          .end((err, res) => {
-            expect(res.body.status).toEqual("error");
-            expect(res.body.message).toEqual("Not Authorised");
-            done();
-          });
+  describe("POST /data/bulk", () => {
+    describe("when invalid", () => {
+      describe("when the token is invalid", () => {
+        it("should be a protected route", done => {
+          request(args.app)
+            .post("/data/bulk")
+            .set("Authorization", "Bearer INVALID")
+            .attach("file", "test/fixtures/experiments-demo/metadata_ena11.tsv")
+            .expect(httpStatus.UNAUTHORIZED)
+            .end((err, res) => {
+              expect(res.body.status).toEqual("error");
+              expect(res.body.message).toEqual("Not Authorised");
+              done();
+            });
+        });
       });
-      it("should check if the folder exists", done => {
-        request(app)
-          .post("/data/demo/experiments-demo-invalid")
-          .set("Authorization", `Bearer ${token}`)
-          .expect(httpStatus.OK)
-          .end((err, res) => {
-            expect(res.body.status).toEqual("error");
-            expect(res.body.message).toEqual(
-              "Cannot find test/fixtures/experiments-demo-invalid directory"
-            );
-            done();
-          });
+      describe("when the file is not a zip", () => {
+        it("should return an error", done => {
+          request(args.app)
+            .post("/data/bulk")
+            .set("Authorization", `Bearer ${args.token}`)
+            .field("purge", true)
+            .attach("file", "test/fixtures/files/333-08.json")
+            .expect(httpStatus.OK)
+            .end((err, res) => {
+              expect(res.body.status).toEqual("error");
+              expect(res.body.message).toEqual("Input file must be a zip file");
+              done();
+            });
+        });
+      });
+      describe("when the zip content is incomplete", () => {
+        it("should return an error", done => {
+          request(args.app)
+            .post("/data/bulk")
+            .set("Authorization", `Bearer ${args.token}`)
+            .field("purge", true)
+            .attach("file", "test/fixtures/files/invalid.zip")
+            .expect(httpStatus.OK)
+            .end((err, res) => {
+              expect(res.body.status).toEqual("error");
+              expect(res.body.message).toEqual("Cannot find metadata files");
+              done();
+            });
+        });
       });
     });
-    describe("when using purge", () => {
-      beforeEach(async done => {
-        const experimentData = new Experiment(experiments.tbUploadMetadata);
-        const experiment = await experimentData.save();
-        done();
-      });
-      it("should purge all the experiments", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", `Bearer ${token}`)
-          .send({ purge: true })
-          .expect(httpStatus.OK)
-          .end(async (err, res) => {
-            let total = await Experiment.count();
-            while (total < 11) {
-              await DataHelper.sleep(100);
-              total = await Experiment.count();
-            }
-            expect(res.body.status).toEqual("success");
-            expect(res.body.data).toEqual(
-              "Demo data upload started from test/fixtures/experiments-demo"
-            );
-            expect(total).toEqual(11);
-            done();
-          });
-      });
-      it("should successfully load the experiments", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", `Bearer ${token}`)
-          .send({ purge: true })
-          .expect(httpStatus.OK)
-          .end(async (err, res) => {
-            let total = await Experiment.count();
-            while (total < 11) {
-              total = await Experiment.count();
-            }
-            expect(res.body.status).toEqual("success");
-            expect(res.body.data).toEqual(
-              "Demo data upload started from test/fixtures/experiments-demo"
-            );
-            const experiment = await Experiment.findOne({
-              "metadata.sample.isolateId": "ERR550945"
-            });
-            expect(experiment.metadata.sample.countryIsolate).toEqual("PG");
-            done();
-          });
-      });
-      it("should replace the city and country", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", `Bearer ${token}`)
-          .send({ purge: true })
-          .expect(httpStatus.OK)
-          .end(async (err, res) => {
-            let total = await Experiment.count();
-            while (total < 11) {
-              total = await Experiment.count();
-            }
-            expect(res.body.status).toEqual("success");
-            expect(res.body.data).toEqual(
-              "Demo data upload started from test/fixtures/experiments-demo"
-            );
-            const experiment = await Experiment.findOne({
-              "metadata.sample.isolateId": "SRR8237379"
-            });
-            expect(experiment.metadata.sample.countryIsolate).toEqual("CH");
-            expect(experiment.metadata.sample.cityIsolate).toEqual("Geneva");
-            done();
-          });
-      });
-      it("should populate geo data", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", `Bearer ${token}`)
-          .send({ purge: true })
-          .expect(httpStatus.OK)
-          .end(async (err, res) => {
-            expect(res.body.status).toEqual("success");
-            expect(res.body.data).toEqual(
-              "Demo data upload started from test/fixtures/experiments-demo"
-            );
-            let experiment = await Experiment.findOne({
-              "metadata.sample.isolateId": "SRR8237379"
-            });
-            while (!experiment || !experiment.metadata.sample.latitudeIsolate) {
-              experiment = await Experiment.findOne({
-                "metadata.sample.isolateId": "SRR8237379"
-              });
-            }
-            expect(experiment.metadata.sample.countryIsolate).toEqual("CH");
-            expect(experiment.metadata.sample.cityIsolate).toEqual("Geneva");
+    describe("when valid", () => {
+      describe("when purge is true", () => {
+        let total = 0;
+        let status = null;
+        let data = null;
 
-            expect(experiment.metadata.sample.latitudeIsolate).toBeGreaterThan(46.2);
-            expect(experiment.metadata.sample.latitudeIsolate).toBeLessThan(47);
+        beforeEach(async done => {
+          const experimentData = new Experiment(experiments.tbUploadMetadata);
+          const experiment = await experimentData.save();
 
-            expect(experiment.metadata.sample.longitudeIsolate).toBeGreaterThan(6.142);
-            expect(experiment.metadata.sample.longitudeIsolate).toBeLessThan(6.147);
-            done();
-          });
-      });
-      it("should not populate the value if unknown", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", `Bearer ${token}`)
-          .send({ purge: true })
-          .expect(httpStatus.OK)
-          .end(async (err, res) => {
-            let total = await Experiment.count();
-            while (total < 11) {
-              total = await Experiment.count();
-            }
-            expect(res.body.status).toEqual("success");
-            expect(res.body.data).toEqual(
-              "Demo data upload started from test/fixtures/experiments-demo"
-            );
-            const experiment = await Experiment.findOne({
-              "metadata.sample.isolateId": "ERR552694"
+          request(args.app)
+            .post("/data/bulk")
+            .set("Authorization", `Bearer ${args.token}`)
+            .field("purge", true)
+            .attach("file", "test/fixtures/files/upload.zip")
+            .expect(httpStatus.OK)
+            .end(async (err, res) => {
+              status = res.body.status;
+              data = res.body.data;
+
+              let total = await Experiment.count();
+              while (total < 3) {
+                await DataHelper.sleep(100);
+                total = await Experiment.count();
+              }
+
+              done();
             });
-            expect(experiment.metadata.sample.countryIsolate).toEqual("");
-            expect(experiment.metadata.sample.cityIsolate).toEqual("");
-            done();
-          });
-      });
-    });
-    describe("when purge is not set", () => {
-      beforeEach(async done => {
-        const experimentData = new Experiment(experiments.tbUploadMetadata);
-        await experimentData.save();
-        done();
-      });
-      it("should not purge the experiments", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", `Bearer ${token}`)
-          .expect(httpStatus.OK)
-          .end(async (err, res) => {
-            let total = await Experiment.count();
-            while (total < 12) {
-              total = await Experiment.count();
+        });
+        it("should return success", done => {
+          expect(status).toEqual("success");
+          done();
+        });
+        it("should trigger loading the experiments", done => {
+          expect(data).toEqual("Demo data upload started");
+          done();
+        });
+        it("should purge all the experiments", async done => {
+          const total = await Experiment.count();
+          expect(total).toEqual(3);
+
+          done();
+        });
+        it("should successfully load the experiments", async done => {
+          const experiments = await Experiment.find({});
+          const results = [];
+
+          for (const experiment of experiments) {
+            expect(experiment).toHaveProperty("metadata");
+            expect(experiment).toHaveProperty("results");
+            expect(experiment).toHaveProperty("id");
+
+            for (const result of experiment.results) {
+              results.push(result);
             }
-            expect(res.body.status).toEqual("success");
-            expect(res.body.data).toEqual(
-              "Demo data upload started from test/fixtures/experiments-demo"
+          }
+
+          for (const result of results) {
+            expect(result).toHaveProperty("type", "predictor");
+          }
+
+          done();
+        });
+        it("should store city and country", async done => {
+          const experiments = await Experiment.find({});
+
+          for (const experiment of experiments) {
+            expect(experiment).toHaveProperty("metadata");
+            expect(experiment.metadata).toHaveProperty("sample");
+            expect(experiment.metadata.sample).toHaveProperty("countryIsolate");
+            expect(experiment.metadata.sample).toHaveProperty("cityIsolate");
+
+            expect(["UK", "AR"]).toContain(experiment.metadata.sample.countryIsolate);
+            expect(["", "Buenos Aires", "Rosario"]).toContain(
+              experiment.metadata.sample.cityIsolate
             );
-            expect(total).toEqual(12);
-            done();
-          });
-      });
-      it("should successfully load the experiments", done => {
-        request(app)
-          .post("/data/demo/experiments-demo")
-          .set("Authorization", `Bearer ${token}`)
-          .expect(httpStatus.OK)
-          .end(async (err, res) => {
-            let total = await Experiment.count();
-            while (total < 12) {
-              total = await Experiment.count();
+          }
+
+          done();
+        });
+        it("should store geo location", async done => {
+          const experiments = await Experiment.find({});
+
+          for (const experiment of experiments) {
+            expect(experiment.metadata.sample).toHaveProperty("latitudeIsolate");
+            expect(experiment.metadata.sample).toHaveProperty("longitudeIsolate");
+
+            const { countryIsolate, cityIsolate } = experiment.metadata.sample;
+            if (countryIsolate === "AR" && cityIsolate === "Rosario") {
+              expect(experiment.metadata.sample.latitudeIsolate).toBeCloseTo(-32.96, 1);
+              expect(experiment.metadata.sample.longitudeIsolate).toBeCloseTo(-60.69, 1);
+            } else if (countryIsolate === "AR" && cityIsolate === "Buenos Aires") {
+              expect(experiment.metadata.sample.latitudeIsolate).toBeCloseTo(-34.61, 1);
+              expect(experiment.metadata.sample.longitudeIsolate).toBeCloseTo(-58.38, 1);
+            } else if (countryIsolate === "UK" && cityIsolate === "") {
+              expect(experiment.metadata.sample.latitudeIsolate).toBeCloseTo(55.37, 1);
+              expect(experiment.metadata.sample.longitudeIsolate).toBeCloseTo(-3.43, 1);
             }
-            expect(res.body.status).toEqual("success");
-            expect(res.body.data).toEqual(
-              "Demo data upload started from test/fixtures/experiments-demo"
-            );
-            const experiment = await Experiment.findOne({
-              "metadata.sample.isolateId": "ERR550945"
+          }
+          done();
+        });
+
+        it("should store predictor results", async done => {
+          const experiments = await Experiment.find({});
+          const results = [];
+
+          for (const experiment of experiments) {
+            for (const result of experiment.results) {
+              results.push(result);
+            }
+          }
+
+          expect(results.length).toEqual(2);
+          for (const result of results) {
+            expect(result).toHaveProperty("type", "predictor");
+          }
+
+          done();
+        });
+      });
+      describe("when purge is false", () => {
+        beforeEach(async done => {
+          const experimentData = new Experiment(experiments.tbUploadMetadata);
+          const experiment = await experimentData.save();
+          request(args.app)
+            .post("/data/bulk")
+            .set("Authorization", `Bearer ${args.token}`)
+            .attach("file", "test/fixtures/files/upload.zip")
+            .expect(httpStatus.OK)
+            .end(async (err, res) => {
+              let total = await Experiment.count();
+              while (total < 4) {
+                await DataHelper.sleep(100);
+                total = await Experiment.count();
+              }
+              done();
             });
-            expect(experiment.metadata.sample.countryIsolate).toEqual("PG");
-            done();
-          });
+        });
+        it("should keep the existing experiments", async done => {
+          const total = await Experiment.count();
+          expect(total).toEqual(4);
+          done();
+        });
       });
     });
   });
