@@ -1,10 +1,11 @@
 export default {
   admin: {
-    baseUrl: "https://accounts.makeandship.com/auth",
+    baseUrl: process.env.KEYCLOAK_URL,
     username: "admin",
-    password: "admin",
+    password: process.env.KEYCLOAK_ADMIN_PASSWORD,
     grantType: "password",
-    clientId: "admin-cli" // use for admin-level operations e.g. register new user
+    clientId: "admin-cli", // use for admin-level operations e.g. register new user
+    realmName: "master"
   },
   client: {
     clientId: "atlas-app", // use for app-level operations e.g. reset password email
