@@ -1,5 +1,4 @@
 import express from "express";
-import logger from "morgan";
 import bodyParser from "body-parser";
 import cookieParser from "cookie-parser";
 import compress from "compression";
@@ -7,25 +6,36 @@ import methodOverride from "method-override";
 import cors from "cors";
 import expressWinston from "express-winston";
 import helmet from "helmet";
-import errors from "errors";
 import httpStatus from "http-status";
 import RateLimit from "express-rate-limit";
 import addRequestId from "express-request-id";
-import winstonInstance from "./modules/winston";
+
+import { ErrorUtil, APIError } from "makeandship-api-common/lib/modules/error";
+import {
+  ExpressInitializer,
+  JsendInitializer
+} from "makeandship-api-common/lib/modules/express/initializers";
+
+import Constants from "./Constants";
+
+import logger from "./modules/logger";
 import routes from "./routes/index.route";
 import config from "../config/env";
-import APIError from "./helpers/APIError";
+
 import AccountsHelper from "./helpers/AccountsHelper";
 import { stubDevApis } from "../external";
 
 const keycloak = AccountsHelper.keycloakInstance();
 
-const createApp = ({ rateLimitReset, rateLimitMax, limit } = config.express) => {
+const createApp = async options => {
+  const settings = Object.assign(config.express, options);
+  const { rateLimitReset, rateLimitMax, limit } = settings;
+
   const app = express();
 
+  // TODO move mocking out
   if (config.env === "development") {
     stubDevApis();
-    app.use(logger("dev"));
   }
 
   // parse body params and attache them to req.body
@@ -57,7 +67,7 @@ const createApp = ({ rateLimitReset, rateLimitMax, limit } = config.express) => 
     expressWinston.responseWhitelist.push("body");
     app.use(
       expressWinston.logger({
-        winstonInstance,
+        winstonInstance: logger,
         meta: true, // optional: log meta data about request (defaults to true)
         msg: "HTTP {{req.method}} {{req.url}} {{res.statusCode}} {{res.responseTime}}ms",
         colorStatus: true, // Color the status code (default green, 3XX cyan, 4XX yellow, 5XX red).
@@ -70,13 +80,15 @@ const createApp = ({ rateLimitReset, rateLimitMax, limit } = config.express) => 
           "query",
           "body"
         ],
+        responseWhitelist: [],
         bodyBlacklist: ["password", "confirmPassword"],
-        ignoreRoute: (req, res) => {
+        ignoreRoute: req => {
           return /health-check/.test(req.url);
         }
       })
     );
   }
+
   // 1000 requests per 15 min
   const limiter = new RateLimit({
     windowMs: rateLimitReset,
@@ -84,7 +96,9 @@ const createApp = ({ rateLimitReset, rateLimitMax, limit } = config.express) => 
     delayMs: 0, // disable delaying - full speed until the max limit is reached
     onLimitReached: (req, res) => {
       throw new APIError(
+        Constants.ERRORS.API_ERROR,
         "Too many requests, please try again later.",
+        null,
         httpStatus.TOO_MANY_REQUESTS
       );
     }
@@ -95,7 +109,12 @@ const createApp = ({ rateLimitReset, rateLimitMax, limit } = config.express) => 
   // catch 404 and forward to error handler
   app.use((req, res, next) => {
     // eslint-disable-line no-unused-vars
-    const err = new APIError("Unknown API route.", httpStatus.NOT_FOUND);
+    const err = new APIError(
+      Constants.ERRORS.ROUTE_NOT_FOUND,
+      "Unknown API route",
+      null,
+      httpStatus.NOT_FOUND
+    );
     return res.jerror(err);
   });
 
@@ -103,6 +122,9 @@ const createApp = ({ rateLimitReset, rateLimitMax, limit } = config.express) => 
   app.use((err, req, res, next) => {
     res.header("Access-Control-Allow-Origin", "*");
     res.header("Access-Control-Allow-Headers", "Origin, X-Requested-With, Content-Type, Accept");
+    if (err.errors) {
+      return res.jerror(ErrorUtil.convert(err, Constants.ERRORS.VALIDATION_ERROR));
+    }
     return res.jerror(err);
   });
 
@@ -110,10 +132,16 @@ const createApp = ({ rateLimitReset, rateLimitMax, limit } = config.express) => 
     // log error in winston transports except when executing test suite
     app.use(
       expressWinston.errorLogger({
-        winstonInstance
+        winstonInstance: logger
       })
     );
   }
+
+  logger.debug("ExpressInitializer#initialize: Initializing ...");
+  const initializer = new ExpressInitializer(express);
+  initializer.add(new JsendInitializer());
+  await initializer.initialize();
+  logger.debug("ExpressInitializer#initialize: Initialization complete");
 
   return app;
 };
