@@ -1,9 +1,9 @@
-import errors from "errors";
 import httpStatus from "http-status";
 import mkdirp from "mkdirp-promise";
 import Promise from "bluebird";
 
-import { ElasticsearchHelper } from "makeandship-api-common/lib/modules/elasticsearch/";
+import { ValidationError, ErrorUtil, APIError } from "makeandship-api-common/lib/modules/error";
+import { ElasticService } from "makeandship-api-common/lib/modules/elasticsearch/";
 import ArrayJSONTransformer from "makeandship-api-common/lib/transformers/ArrayJSONTransformer";
 import SearchResultsJSONTransformer from "makeandship-api-common/lib/modules/elasticsearch/transformers/SearchResultsJSONTransformer";
 import SearchQueryJSONTransformer from "makeandship-api-common/lib/modules/elasticsearch/transformers/SearchQueryJSONTransformer";
@@ -19,16 +19,20 @@ import Audit from "../models/audit.model";
 import Experiment from "../models/experiment.model";
 import Tree from "../models/tree.model";
 
+import SearchQueryDecorator from "../modules/search/search-query-decorator";
+import RequestSearchQueryParser from "../modules/search/request-search-query-parser";
+
 import resumable from "../modules/resumable";
 import { schedule } from "../modules/agenda";
 import { experimentEventEmitter, userEventEmitter } from "../modules/events";
 import { parseQuery, callTreeApi } from "../modules/search";
-import winston from "../modules/winston";
+import logger from "../modules/logger";
 
-import APIError from "../helpers/APIError";
 import DownloadersFactory from "../helpers/DownloadersFactory";
 import BigsiSearchHelper from "../helpers/BigsiSearchHelper";
 import ResultsParserFactory from "../helpers/results/ResultsParserFactory";
+import EventHelper from "../helpers/events/EventHelper";
+import EventProgress from "../helpers/events/EventProgress";
 
 import AuditJSONTransformer from "../transformers/AuditJSONTransformer";
 import ExperimentJSONTransformer from "../transformers/ExperimentJSONTransformer";
@@ -39,8 +43,8 @@ import config from "../../config/env";
 import Constants from "../Constants";
 import SearchJSONTransformer from "../transformers/SearchJSONTransformer";
 
-// sort whitelist
-const sortWhitelist = ElasticsearchHelper.getSortWhitelist(experimentSearchSchema, "experiment");
+const esConfig = { type: "experiment", ...config.elasticsearch };
+const elasticService = new ElasticService(esConfig, experimentSearchSchema);
 
 // distance types
 const NEAREST_NEIGHBOUR = "nearest-neighbour";
@@ -95,10 +99,10 @@ const create = async (req, res) => {
 
   try {
     const savedExperiment = await experiment.save();
-    await ElasticsearchHelper.indexDocument(config, savedExperiment, "experiment");
+    await elasticService.indexDocument(savedExperiment);
     return res.jsend(savedExperiment);
   } catch (e) {
-    return res.jerror(new errors.CreateExperimentError(e.message));
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.CREATE_EXPERIMENT));
   }
 };
 
@@ -117,10 +121,10 @@ const update = async (req, res) => {
 
   try {
     const savedExperiment = await experiment.save();
-    await ElasticsearchHelper.updateDocument(config, savedExperiment, "experiment");
+    await elasticService.updateDocument(savedExperiment);
     return res.jsend(savedExperiment);
   } catch (e) {
-    return res.jerror(new errors.UpdateExperimentError(e.message));
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.UPDATE_EXPERIMENT));
   }
 };
 
@@ -150,10 +154,10 @@ const remove = async (req, res) => {
   const experiment = req.experiment;
   try {
     await experiment.remove();
-    await ElasticsearchHelper.deleteDocument(config, experiment.id, "experiment");
+    await elasticService.deleteDocument(experiment.id);
     return res.jsend("Experiment was successfully deleted.");
   } catch (e) {
-    return res.jerror(e);
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.DELETE_EXPERIMENT));
   }
 };
 
@@ -168,10 +172,10 @@ const metadata = async (req, res) => {
 
   try {
     const savedExperiment = await experiment.save();
-    await ElasticsearchHelper.updateDocument(config, savedExperiment, "experiment");
+    await elasticService.updateDocument(savedExperiment);
     return res.jsend(savedExperiment);
   } catch (e) {
-    return res.jerror(new errors.UpdateExperimentError(e.message));
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.UPDATE_EXPERIMENT));
   }
 };
 
@@ -185,7 +189,9 @@ const results = async (req, res) => {
 
   const parser = await ResultsParserFactory.create(req.body);
   if (!parser) {
-    return res.jerror(new errors.UpdateExperimentError("Invalid result type."));
+    return res.jerror(
+      new APIError(Constants.ERRORS.UPDATE_EXPERIMENT_RESULTS, "Invalid result type")
+    );
   }
 
   const result = parser.parse(req.body);
@@ -197,29 +203,31 @@ const results = async (req, res) => {
   }
 
   updatedResults.push(result);
-
   experiment.set("results", updatedResults);
 
   try {
     const savedExperiment = await experiment.save();
 
-    await ElasticsearchHelper.updateDocument(config, savedExperiment, "experiment");
+    await elasticService.updateDocument(savedExperiment);
 
     const audit = await Audit.getByExperimentId(savedExperiment.id);
 
     const experimentJSON = new ExperimentJSONTransformer().transform(experiment);
     const auditJSON = audit ? new AuditJSONTransformer().transform(audit) : null;
 
+    await EventHelper.clearAnalysisState(savedExperiment.id);
+
     experimentEventEmitter.emit("analysis-complete", {
       audit: auditJSON,
       experiment: experimentJSON,
       type: result.type,
-      subType: result.subType
+      subType: result.subType,
+      fileLocation: result.files
     });
 
     return res.jsend(savedExperiment);
   } catch (e) {
-    return res.jerror(new errors.UpdateExperimentError(e.message));
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.UPDATE_EXPERIMENT_RESULTS));
   }
 };
 
@@ -237,13 +245,17 @@ const uploadFile = async (req, res) => {
       await mkdirp(path);
       const downloader = DownloadersFactory.create(`${path}/${req.body.name}`, {
         experiment,
+        user: req.dbUser,
         ...req.body
       });
       downloader.download(async () => {
+        await EventHelper.updateAnalysisState(
+          req.dbUser.id,
+          experiment.id,
+          `${config.express.uploadsLocation}/experiments/${experiment.id}/file/${req.body.name}`
+        );
         await schedule("now", "call analysis api", {
-          file: `${config.express.uploadsLocation}/experiments/${experiment.id}/file/${
-            req.body.name
-          }`,
+          file: `${config.express.uploadsLocation}/experiments/${experiment.id}/file/${req.body.name}`,
           experiment_id: experiment.id,
           attempt: 0
         });
@@ -254,39 +266,58 @@ const uploadFile = async (req, res) => {
 
       return res.jsend(`Download started from ${req.body.provider}`);
     } catch (e) {
-      return res.jerror(new errors.UploadFileError(e.message));
+      return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.UPLOAD_FILE));
     }
   }
 
   // no file provided to upload
   if (!req.file) {
-    return res.jerror(new errors.UploadFileError("No files found to upload"));
+    return res.jerror(new APIError(Constants.ERRORS.UPLOAD_FILE, "No files found to upload"));
   }
 
   // from local file
   try {
     const experimentJson = new ExperimentJSONTransformer().transform(req.experiment);
     const resumableFilename = req.body.resumableFilename;
-
-    await resumable.setUploadDirectory(
-      `${config.express.uploadDir}/experiments/${experiment.id}/file`
-    );
+    const uploadDirectory = `${config.express.uploadDir}/experiments/${experiment.id}/file`;
+    logger.debug(`ExperimentsController#uploadFile: uploadDirectory: ${uploadDirectory}`);
+    await resumable.setUploadDirectory(uploadDirectory);
     const postUpload = await resumable.post(req);
     if (!postUpload.complete) {
-      experimentEventEmitter.emit("upload-progress", {
-        experiment: experimentJson,
-        status: postUpload
-      });
+      logger.debug(`ExperimentsController#uploadFile: more`);
+      const currentProgress = EventProgress.get(postUpload);
+      // only update progress for each percent change
+      const diff = EventProgress.diff(postUpload.id, postUpload);
+      logger.debug(`ExperimentsController#uploadFile: diff: ${JSON.stringify(diff, null, 2)}`);
+      if (diff > 1 || !currentProgress) {
+        try {
+          await EventHelper.updateUploadsState(req.dbUser.id, experiment.id, postUpload);
+        } catch (e) {
+          logger.error(`Unable to store upload state: ${JSON.stringify(e, null, 2)}`);
+        }
+        experimentEventEmitter.emit("upload-progress", {
+          experiment: experimentJson,
+          status: postUpload
+        });
+        EventProgress.update(postUpload.id, postUpload);
+      }
     } else {
+      logger.debug(`ExperimentsController#uploadFile: complete`);
+      await EventHelper.clearUploadsState(req.dbUser.id, experiment.id);
       experimentEventEmitter.emit("upload-complete", {
         experiment: experimentJson,
         status: postUpload
       });
+      logger.debug(`ExperimentsController#uploadFile: updateAnalysisState`);
+      await EventHelper.updateAnalysisState(
+        req.dbUser.id,
+        experimentJson.id,
+        `${config.express.uploadsLocation}/experiments/${experimentJson.id}/file/${resumableFilename}`
+      );
+      logger.debug(`ExperimentsController#uploadFile: reassembleChunks ...`);
       return resumable.reassembleChunks(experimentJson.id, resumableFilename, async () => {
         await schedule("now", "call analysis api", {
-          file: `${config.express.uploadsLocation}/experiments/${
-            experimentJson.id
-          }/file/${resumableFilename}`,
+          file: `${config.express.uploadsLocation}/experiments/${experimentJson.id}/file/${resumableFilename}`,
           experiment_id: experimentJson.id,
           attempt: 0,
           experiment: experimentJson
@@ -294,9 +325,11 @@ const uploadFile = async (req, res) => {
         return res.jsend("File uploaded and reassembled");
       });
     }
-    return res.jerror(postUpload);
-  } catch (err) {
-    return res.jerror(new errors.UploadFileError(err.message));
+    return res.jerror(
+      new APIError(Constants.ERRORS.UPLOAD_FILE, "Error uploading file", postUpload)
+    );
+  } catch (e) {
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.UPLOAD_FILE));
   }
 };
 
@@ -309,24 +342,34 @@ const readFile = (req, res) => {
     const path = `${config.express.uploadDir}/experiments/${experiment.id}/file`;
     return res.sendFile(`${path}/${experiment.file}`);
   }
-  return res.jerror("No file found for this Experiment");
+
+  return res.jerror(new APIError(Constants.ERRORS.GET_EXPERIMENT, "Error reading file"));
 };
 
 const uploadStatus = async (req, res) => {
+  logger.debug(`ExperimentController#uploadStatus: enter`);
   const experiment = req.experiment;
   try {
-    await resumable.setUploadDirectory(
-      `${config.express.uploadDir}/experiments/${experiment.id}/file`
-    );
+    const uploadDirectory = `${config.express.uploadDir}/experiments/${experiment.id}/file`;
+    logger.debug(`ExperimentController#uploadStatus: uploadDirectory: ${uploadDirectory}`);
+    await resumable.setUploadDirectory(uploadDirectory);
 
     const validateGetRequest = resumable.get(req);
+    logger.debug(
+      `ExperimentController#uploadStatus: validateGetRequest: ${JSON.stringify(validateGetRequest)}`
+    );
     if (validateGetRequest.valid) {
       return res.jsend(validateGetRequest);
     }
-    const error = new APIError(validateGetRequest.message, httpStatus.NO_CONTENT);
+    const error = new APIError(
+      Constants.ERRORS.UPLOAD_FILE,
+      validateGetRequest.message,
+      null,
+      httpStatus.NO_CONTENT
+    );
     return res.jerror(error);
-  } catch (err) {
-    return res.jerror(new errors.UploadFileError(err.message));
+  } catch (e) {
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.UPLOAD_FILE));
   }
 };
 
@@ -335,14 +378,32 @@ const uploadStatus = async (req, res) => {
  */
 const reindex = async (req, res) => {
   try {
-    const total = await ElasticsearchHelper.reindex(config, experimentSearchSchema, {
-      type: "experiment",
-      model: Experiment,
-      size: req.body.size || req.query.size
-    });
-    return res.jsend(`All ${total} experiment(s) have been indexed.`);
+    const { indexSizeLimit } = config.elasticsearch;
+    const size = req.body.size || req.query.size || indexSizeLimit;
+
+    await elasticService.deleteIndex();
+    await elasticService.createIndex();
+    // index in batches
+    const pagination = {
+      count: 0,
+      more: true,
+      id: null
+    };
+    while (pagination.more) {
+      const data = await Experiment.since(pagination.id, parseInt(size));
+      const result = await elasticService.indexDocuments(data);
+
+      if (data.length === parseInt(size)) {
+        pagination.more = true;
+        pagination.id = data[data.length - 1]._id;
+      } else {
+        pagination.more = false;
+      }
+      pagination.count = pagination.count + data.length;
+    }
+    return res.jsend(`All ${pagination.count} experiment(s) have been indexed.`);
   } catch (e) {
-    return res.jerror(e.message);
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.REINDEX_EXPERIMENTS));
   }
 };
 
@@ -355,23 +416,25 @@ const choices = async (req, res) => {
     const container = parseQuery(clone);
     const query = container.query;
 
-    const resp = await ElasticsearchHelper.aggregate(
-      config,
-      experimentSearchSchema,
-      query,
-      "experiment"
-    );
+    // parse the query
+    const parsedQuery = new RequestSearchQueryParser(req.originalUrl).parse(query);
+
+    // apply status and organisation filters
+    const searchQuery = new SearchQueryDecorator(req.originalUrl, req.user).decorate(parsedQuery);
+    const elasticsearchResults = await elasticService.search(searchQuery, {});
+
     const titles = jsonschemaUtil.schemaTitles(experimentSearchSchema);
-    const choices = new ChoicesJSONTransformer().transform(resp, { titles });
+
+    const choices = await new ChoicesJSONTransformer().transform(elasticsearchResults, { titles });
 
     return res.jsend(choices);
   } catch (e) {
-    return res.jerror(new errors.SearchMetadataValuesError(e.message));
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.SEARCH_METADATA_VALUES));
   }
 };
 
 /**
- * Get experiments list from ES.
+ * Get experiment list.
  * @returns {Experiment[]}
  */
 const search = async (req, res) => {
@@ -385,33 +448,37 @@ const search = async (req, res) => {
     if (bigsi) {
       const search = await BigsiSearchHelper.search(bigsi, query, req.dbUser);
       const searchJson = new SearchJSONTransformer().transform(search);
-
       searchJson.search = new SearchQueryJSONTransformer().transform(req.query, {});
 
       return res.jsend(searchJson);
     } else {
-      clone.whitelist = sortWhitelist;
+      // parse the query
+      const parsedQuery = new RequestSearchQueryParser(req.originalUrl).parse(clone);
 
-      const resp = await ElasticsearchHelper.search(config, clone, "experiment");
+      // apply status and organisation filters
+      const searchQuery = new SearchQueryDecorator(req.originalUrl, req.user).decorate(parsedQuery);
+      const elasticsearchResults = await elasticService.search(searchQuery, {});
 
       // generate the core elastic search structure
       const options = {
-        per: query.per || config.elasticsearch.resultsPerPage,
-        page: query.page || 1
+        per: req.query.per || config.elasticsearch.resultsPerPage,
+        page: req.query.page || 1
       };
+      const results = new SearchResultsJSONTransformer().transform(elasticsearchResults, options);
 
-      const results = new SearchResultsJSONTransformer().transform(resp, options);
       if (results) {
         // augment with hits (project specific transformation)
-        results.results = new ExperimentsResultJSONTransformer().transform(resp, {});
-
+        results.results = new ExperimentsResultJSONTransformer().transform(
+          elasticsearchResults,
+          {}
+        );
         // augment with the original search query
-        results.search = new SearchQueryJSONTransformer().transform(req.query, {});
+        results.search = new SearchQueryJSONTransformer().transform(searchQuery, {});
       }
       return res.jsend(results);
     }
   } catch (e) {
-    return res.jerror(new errors.SearchMetadataValuesError(e.message));
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.SEARCH_METADATA_VALUES));
   }
 };
 
@@ -498,6 +565,16 @@ const refreshResults = async (req, res) => {
   return res.jsend("Update of existing results triggered");
 };
 
+const mappings = async (req, res) => {
+  const experiments = await Experiment.list();
+  const result = {};
+  experiments.forEach(experiment => {
+    const metadata = experiment.get("metadata");
+    result[metadata.sample.isolateId] = experiment.id;
+  });
+  return res.jsend(result);
+};
+
 export default {
   load,
   get,
@@ -515,5 +592,6 @@ export default {
   search,
   listResults,
   tree,
-  refreshResults
+  refreshResults,
+  mappings
 };
