@@ -1,30 +1,23 @@
 import passwordHash from "password-hash";
-import errors from "errors";
 import flatten from "flat";
 
-import { ElasticsearchHelper } from "makeandship-api-common/lib/modules/elasticsearch/";
+import { ErrorUtil, APIError } from "makeandship-api-common/lib/modules/error";
 import ArrayJSONTransformer from "makeandship-api-common/lib/transformers/ArrayJSONTransformer";
 
-import { userEventEmitter } from "../modules/events";
 import channels from "../modules/channels";
 
-import Audit from "../models/audit.model";
-import Experiment from "../models/experiment.model";
-import Search from "../models/search.model";
 import User from "../models/user.model";
+import Event from "../models/event.model";
 
-import AuditJSONTransformer from "../transformers/AuditJSONTransformer";
 import UserJSONTransformer from "../transformers/UserJSONTransformer";
-import ExperimentJSONTransformer from "../transformers/ExperimentJSONTransformer";
-import ExperimentsResultJSONTransformer from "../transformers/es/ExperimentsResultJSONTransformer";
+import EventJSONTransformer from "../transformers/EventJSONTransformer";
 
 import AccountsHelper from "../helpers/AccountsHelper";
 import EmailHelper from "../helpers/EmailHelper";
 
-import ResultsParserFactory from "../helpers/results/ResultsParserFactory";
-
 import config from "../../config/env";
-import SearchJSONTransformer from "../transformers/SearchJSONTransformer";
+
+import Constants from "../Constants";
 
 const keycloak = AccountsHelper.keycloakInstance();
 
@@ -70,7 +63,17 @@ const create = async (req, res) => {
 
     if (EmailHelper.isValid(username)) {
       if (!req.body.password) {
-        return res.jerror("Please provide a password");
+        return res.jerror(
+          new APIError(Constants.ERRORS.CREATE_USER, "Please provide a password", {
+            errors: {
+              password: {
+                path: "password",
+                type: "required",
+                message: "should have required property 'password'"
+              }
+            }
+          })
+        );
       }
 
       userData.email = username;
@@ -86,14 +89,24 @@ const create = async (req, res) => {
         req.body.password
       );
     } else {
-      return res.jerror("Invalid username");
+      return res.jerror(
+        new APIError(Constants.ERRORS.CREATE_USER, "Invalid username", {
+          errors: {
+            username: {
+              path: "username",
+              type: "required",
+              message: "should be a valid username"
+            }
+          }
+        })
+      );
     }
 
     const user = new User(userData);
     const savedUser = await user.save();
     return res.jsend(savedUser);
   } catch (e) {
-    return res.jerror(e);
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.CREATE_USER));
   }
 };
 
@@ -105,6 +118,7 @@ const create = async (req, res) => {
  */
 const update = async (req, res) => {
   const user = req.dbUser;
+
   user.firstname = req.body.firstname || user.firstname;
   user.lastname = req.body.lastname || user.lastname;
   user.phone = typeof req.body.phone === "undefined" ? user.phone : req.body.phone;
@@ -114,7 +128,7 @@ const update = async (req, res) => {
     const savedUser = await user.save({ lean: true });
     return res.jsend(new UserJSONTransformer().transform(savedUser));
   } catch (e) {
-    return res.jerror(e);
+    return res.jerror(ErrorUtil.convert(e, Constants.ERRORS.UPDATE_USER));
   }
 };
 
@@ -179,6 +193,19 @@ const events = async (req, res) => {
   channel.addClient(req, res);
 };
 
+/**
+ * Get user events status
+ */
+const eventsStatus = async (req, res) => {
+  const user = req.dbUser;
+  try {
+    const event = await Event.getByUserId(user.id);
+    return res.jsend(new EventJSONTransformer().transform(event));
+  } catch (e) {
+    return res.jerror(e);
+  }
+};
+
 export default {
   load,
   get,
@@ -188,5 +215,6 @@ export default {
   remove,
   assignRole,
   loadCurrentUser,
-  events
+  events,
+  eventsStatus
 };
