@@ -1,9 +1,10 @@
-import errors from "errors";
+import httpStatus from "http-status";
 import { Keycloak } from "makeandship-api-common/lib/modules/accounts";
+import { AuthError } from "makeandship-api-common/lib/modules/error";
 import config from "../../config/env";
 import User from "../models/user.model";
 import UserJSONTransformer from "../transformers/UserJSONTransformer";
-import APIError from "../helpers/APIError";
+import Constants from "../Constants";
 
 class AccountsHelper {
   static usePassword(config) {
@@ -13,11 +14,26 @@ class AccountsHelper {
   static keycloakInstance() {
     return new Keycloak({
       keycloakConfig: config.accounts.keycloak,
-      errors,
       userModel: User,
       userTransformer: UserJSONTransformer,
-      apiError: APIError
+      apiError: new AuthError(
+        Constants.ERRORS.INVALID_CREDENTIALS,
+        null,
+        null,
+        httpStatus.UNAUTHORIZED
+      )
     });
+  }
+
+  static async setupGroupsAndRoles(organisation) {
+    const keycloak = this.keycloakInstance();
+    const membersGroup = await keycloak.createGroupIfMissing(`${organisation.slug}-members`);
+    const ownersGroup = await keycloak.createGroupIfMissing(`${organisation.slug}-owners`);
+    const roleName = await keycloak.createRoleIfMissing(organisation.slug);
+    organisation.membersGroupId = membersGroup.id;
+    organisation.ownersGroupId = ownersGroup.id;
+    await keycloak.createGroupRoleMapping(organisation.membersGroupId, roleName);
+    await keycloak.createGroupRoleMapping(organisation.ownersGroupId, roleName);
   }
 }
 
