@@ -1,7 +1,9 @@
 import https from "https";
 import fs from "fs";
-import winston from "winston";
+import logger from "../modules/logger";
 import { experimentEventEmitter } from "../modules/events";
+import EventHelper from "../helpers/events/EventHelper";
+import EventProgress from "../helpers/events/EventProgress";
 
 /**
  * A class to download large files from a url
@@ -15,49 +17,71 @@ class Downloader {
   }
 
   download(done) {
-    const file = fs.createWriteStream(this.destination);
+    const that = this;
+
+    const experiment = that.data.experiment;
+    const provider = that.data.provider;
+    const path = that.options.path;
+    const user = that.data.user;
+    const userId = user ? user.id : null;
+
+    const file = fs.createWriteStream(that.destination);
+
     https.get(this.options).on("response", res => {
       let downloaded = 0;
       const totalSize = res.headers["content-length"];
       res
-        .on("data", chunk => {
+        .on("data", async chunk => {
           file.write(chunk);
           downloaded += chunk.length;
-
-          const experiment = this.data.experiment;
           const status = {
-            provider: this.data.provider,
+            id: experiment.id,
+            provider: provider,
             size: downloaded,
             totalSize,
-            fileLocation: this.options.path
+            fileLocation: path
           };
-
-          experimentEventEmitter.emit("3rd-party-upload-progress", {
-            experiment,
-            status
-          });
+          const diff = EventProgress.diff(experiment.id, status);
+          if (diff > 1) {
+            EventProgress.update(experiment.id, status);
+            try {
+              await EventHelper.updateUploadsState(userId, experiment.id, status);
+            } catch (e) {
+              logger.error(`Error updating uploads state: ${JSON.stringify(e, null, 2)}`);
+            }
+            experimentEventEmitter.emit("3rd-party-upload-progress", {
+              experiment,
+              status
+            });
+          }
         })
-        .on("end", () => {
+        .on("end", async () => {
+          file.end();
           if (done) {
             done();
           }
-          file.end();
-
-          const experiment = this.data.experiment;
           const status = {
-            provider: this.data.provider,
+            provider: provider,
             size: totalSize,
             totalSize,
-            fileLocation: this.options.path
+            fileLocation: path
           };
 
+          try {
+            await EventHelper.clearUploadsState(this.data.user.id, experiment.id);
+          } catch (e) {
+            logger.error(`Error clearing uploads state: ${JSON.stringify(e, null, 2)}`);
+          }
           experimentEventEmitter.emit("3rd-party-upload-complete", {
             experiment,
             status
           });
         })
         .on("error", err => {
-          winston.info(err.message);
+          done(err.message);
+        })
+        .on("aborted", err => {
+          done(err);
         });
     });
   }
