@@ -1,8 +1,8 @@
 import EventEmitter from "events";
-import Experiment from "../models/experiment.model";
-import Audit from "../models/audit.model";
 
 import channels from "./channels";
+
+import Constants from "../Constants";
 
 import UploadProgressJSONTransformer from "../transformers/events/UploadProgressJSONTransformer";
 import UploadCompleteJSONTransformer from "../transformers/events/UploadCompleteJSONTransformer";
@@ -15,11 +15,16 @@ import SequenceSearchStartedEventJSONTransformer from "../transformers/events/Se
 import SequenceSearchCompleteEventJSONTransformer from "../transformers/events/SequenceSearchCompleteEventJSONTransformer";
 import ProteinVariantSearchStartedEventJSONTransformer from "../transformers/events/ProteinVariantSearchStartedEventJSONTransformer";
 import ProteinVariantSearchCompleteEventJSONTransformer from "../transformers/events/ProteinVariantSearchCompleteEventJSONTransformer";
+import DnaVariantSearchStartedEventJSONTransformer from "../transformers/events/DnaVariantSearchStartedEventJSONTransformer";
+import DnaVariantSearchCompleteEventJSONTransformer from "../transformers/events/DnaVariantSearchCompleteEventJSONTransformer";
+
+import logger from "./logger";
 
 const experimentEventEmitter = new EventEmitter();
 const userEventEmitter = new EventEmitter();
 
 const sendUserEvent = (userId, data) => {
+  logger.debug(`sendUserEvent: userId: ${userId}`);
   if (userId) {
     const channel = channels.getUserChannel(userId);
     channel.send({ data });
@@ -28,6 +33,7 @@ const sendUserEvent = (userId, data) => {
 
 const sendExperimentOwnerEvent = (experiment, data, type) => {
   const owner = experiment.owner;
+  logger.debug(`sendExperimentOwnerEvent: owner: ${owner}`);
   if (owner) {
     const ownerId = owner.id || owner;
     if (ownerId) {
@@ -36,25 +42,25 @@ const sendExperimentOwnerEvent = (experiment, data, type) => {
   }
 };
 
-experimentEventEmitter.on("upload-progress", payload => {
+experimentEventEmitter.on(Constants.EVENTS.UPLOAD_PROGRESS.EVENT, payload => {
   const { experiment, status } = payload;
 
   if (experiment && status) {
     const data = new UploadProgressJSONTransformer().transform({ experiment, status }, {});
-    sendExperimentOwnerEvent(experiment, data);
+    sendExperimentOwnerEvent(experiment, data, Constants.EVENTS.UPLOAD_PROGRESS.EVENT);
   }
 });
 
-experimentEventEmitter.on("upload-complete", payload => {
+experimentEventEmitter.on(Constants.EVENTS.UPLOAD_COMPLETE.EVENT, payload => {
   const { experiment, status } = payload;
 
   if (experiment && status) {
     const data = new UploadCompleteJSONTransformer().transform({ experiment, status }, {});
-    sendExperimentOwnerEvent(experiment, data);
+    sendExperimentOwnerEvent(experiment, data, Constants.EVENTS.UPLOAD_COMPLETE.EVENT);
   }
 });
 
-experimentEventEmitter.on("3rd-party-upload-progress", payload => {
+experimentEventEmitter.on(Constants.EVENTS.THIRD_PARTY_UPLOAD_PROGRESS.EVENT, payload => {
   const { experiment, status } = payload;
 
   if (experiment && status) {
@@ -62,11 +68,11 @@ experimentEventEmitter.on("3rd-party-upload-progress", payload => {
       { experiment, status },
       {}
     );
-    sendExperimentOwnerEvent(experiment, data);
+    sendExperimentOwnerEvent(experiment, data, Constants.EVENTS.THIRD_PARTY_UPLOAD_PROGRESS.EVENT);
   }
 });
 
-experimentEventEmitter.on("3rd-party-upload-complete", payload => {
+experimentEventEmitter.on(Constants.EVENTS.THIRD_PARTY_UPLOAD_COMPLETE.EVENT, payload => {
   const { experiment, status } = payload;
 
   if (experiment && status) {
@@ -74,40 +80,42 @@ experimentEventEmitter.on("3rd-party-upload-complete", payload => {
       { experiment, status },
       {}
     );
-    sendExperimentOwnerEvent(experiment, data);
+    sendExperimentOwnerEvent(experiment, data, Constants.EVENTS.THIRD_PARTY_UPLOAD_COMPLETE.EVENT);
   }
 });
 
-experimentEventEmitter.on("analysis-started", async audit => {
+experimentEventEmitter.on(Constants.EVENTS.ANALYSIS_STARTED.EVENT, async payload => {
   try {
     const { experiment, audit } = payload;
 
     if (audit && experiment) {
-      const data = new AnalysisStartedJSONTransformer().transform({ audit, experiment }, {});
-      sendExperimentOwnerEvent(experiment, data);
-    }
-  } catch (e) {}
-});
-
-experimentEventEmitter.on("analysis-complete", async payload => {
-  try {
-    const { experiment, type, audit } = payload;
-
-    if (experiment && type && audit) {
-      const data = new AnalysisCompleteJSONTransformer().transform(
-        {
-          audit,
-          experiment,
-          type
-        },
+      const data = new AnalysisStartedJSONTransformer().transform(
+        { audit, experiment, fileLocation: audit.fileLocation },
         {}
       );
-      sendExperimentOwnerEvent(payload.experiment, data);
+      sendExperimentOwnerEvent(experiment, data, Constants.EVENTS.ANALYSIS_STARTED.EVENT);
     }
-  } catch (e) {}
+  } catch (e) {
+    logger.error(`analysis-started: error: ${JSON.stringify(e, null, 2)}`);
+  }
 });
 
-experimentEventEmitter.on("distance-search-started", async payload => {
+experimentEventEmitter.on(Constants.EVENTS.ANALYSIS_COMPLETE.EVENT, async payload => {
+  try {
+    const { experiment, type, audit } = payload;
+    logger.debug(`Analysis complete`);
+    if (experiment && type && audit) {
+      logger.debug(`Analysis complete data valid: ${JSON.stringify(payload, null, 2)}`);
+      const data = new AnalysisCompleteJSONTransformer().transform(payload, {});
+      logger.debug(`Analysis complete transformed data: ${JSON.stringify(data, null, 2)}`);
+      sendExperimentOwnerEvent(payload.experiment, data, Constants.EVENTS.ANALYSIS_COMPLETE.EVENT);
+    }
+  } catch (e) {
+    logger.error(`analysis-completed: error: ${JSON.stringify(e, null, 2)}`);
+  }
+});
+
+experimentEventEmitter.on(Constants.EVENTS.DISTANCE_SEARCH_STARTED.EVENT, async payload => {
   try {
     const { experiment, audit } = payload;
 
@@ -119,15 +127,15 @@ experimentEventEmitter.on("distance-search-started", async payload => {
         },
         {}
       );
-      sendExperimentOwnerEvent(experiment, data);
+      sendExperimentOwnerEvent(experiment, data, Constants.EVENTS.DISTANCE_SEARCH_STARTED.EVENT);
     }
   } catch (e) {}
 });
 
-userEventEmitter.on("sequence-search-started", async payload => {
+userEventEmitter.on(Constants.EVENTS.SEQUENCE_SEARCH_STARTED.EVENT, async payload => {
   try {
+    logger.debug(`Sequence search started`);
     const { audit, search, user } = payload;
-
     if (audit && search && user) {
       const data = new SequenceSearchStartedEventJSONTransformer().transform(
         {
@@ -139,15 +147,16 @@ userEventEmitter.on("sequence-search-started", async payload => {
       );
 
       const userId = user.id;
+      logger.debug(`Sequence search started: Send event`);
       sendUserEvent(userId, data);
     }
   } catch (e) {}
 });
 
-userEventEmitter.on("sequence-search-complete", async payload => {
+userEventEmitter.on(Constants.EVENTS.SEQUENCE_SEARCH_COMPLETE.EVENT, async payload => {
   try {
+    logger.debug(`Sequence search complete`);
     const { audit, search, user } = payload;
-
     if (audit && search && user) {
       const data = new SequenceSearchCompleteEventJSONTransformer().transform(
         {
@@ -159,14 +168,20 @@ userEventEmitter.on("sequence-search-complete", async payload => {
       );
 
       const userId = user.id;
+      logger.debug(`Sequence search complete: Send event`);
       sendUserEvent(userId, data);
     }
   } catch (e) {}
 });
 
-userEventEmitter.on("protein-variant-search-started", async payload => {
+userEventEmitter.on(Constants.EVENTS.PROTEIN_VARIANT_SEARCH_STARTED.EVENT, async payload => {
   try {
+    logger.debug(`Protein variant search started`);
     const { search, audit, user } = payload;
+
+    logger.debug(`userEventEmitter: audit: ${JSON.stringify(audit, null, 2)}`);
+    logger.debug(`userEventEmitter: user: ${JSON.stringify(user, null, 2)}`);
+    logger.debug(`userEventEmitter: search: ${JSON.stringify(search, null, 2)}`);
 
     if (audit && search && user) {
       const data = new ProteinVariantSearchStartedEventJSONTransformer().transform(
@@ -179,14 +194,20 @@ userEventEmitter.on("protein-variant-search-started", async payload => {
       );
 
       const userId = user.id;
+      logger.debug(`Protein variant search started: Send event`);
       sendUserEvent(userId, data);
     }
   } catch (e) {}
 });
 
-userEventEmitter.on("protein-variant-search-complete", async payload => {
+userEventEmitter.on(Constants.EVENTS.PROTEIN_VARIANT_SEARCH_COMPLETE.EVENT, async payload => {
   try {
+    logger.debug(`Protein variant search complete`);
     const { search, audit, user } = payload;
+
+    logger.debug(`userEventEmitter: audit: ${JSON.stringify(audit, null, 2)}`);
+    logger.debug(`userEventEmitter: user: ${JSON.stringify(user, null, 2)}`);
+    logger.debug(`userEventEmitter: search: ${JSON.stringify(search, null, 2)}`);
 
     if (audit && search && user) {
       const data = new ProteinVariantSearchCompleteEventJSONTransformer().transform(
@@ -198,14 +219,20 @@ userEventEmitter.on("protein-variant-search-complete", async payload => {
         {}
       );
       const userId = user.id;
+      logger.debug(`Protein variant search complete: Send event`);
       sendUserEvent(userId, data);
     }
   } catch (e) {}
 });
 
-userEventEmitter.on("dna-variant-search-complete", async audit => {
+userEventEmitter.on(Constants.EVENTS.DNA_VARIANT_SEARCH_STARTED.EVENT, async payload => {
   try {
-    const { experiment, audit, user } = payload;
+    logger.debug(`DNA variant search started`);
+    const { search, audit, user } = payload;
+
+    logger.debug(`userEventEmitter: audit: ${JSON.stringify(audit, null, 2)}`);
+    logger.debug(`userEventEmitter: user: ${JSON.stringify(user, null, 2)}`);
+    logger.debug(`userEventEmitter: search: ${JSON.stringify(search, null, 2)}`);
 
     if (audit && search && user) {
       const data = new DnaVariantSearchStartedJSONTransformer().transform(
@@ -217,6 +244,32 @@ userEventEmitter.on("dna-variant-search-complete", async audit => {
         {}
       );
       const userId = user.id;
+      logger.debug(`DNA variant search started: Send event`);
+      sendUserEvent(userId, data);
+    }
+  } catch (e) {}
+});
+
+userEventEmitter.on(Constants.EVENTS.DNA_VARIANT_SEARCH_COMPLETE.EVENT, async payload => {
+  try {
+    logger.debug(`DNA variant search complete`);
+    const { search, audit, user } = payload;
+
+    logger.debug(`userEventEmitter: audit: ${JSON.stringify(audit, null, 2)}`);
+    logger.debug(`userEventEmitter: user: ${JSON.stringify(user, null, 2)}`);
+    logger.debug(`userEventEmitter: search: ${JSON.stringify(search, null, 2)}`);
+
+    if (audit && search && user) {
+      const data = new DnaVariantSearchCompletedJSONTransformer().transform(
+        {
+          audit,
+          search,
+          user
+        },
+        {}
+      );
+      const userId = user.id;
+      logger.debug(`DNA variant search complete: Send event`);
       sendUserEvent(userId, data);
     }
   } catch (e) {}
