@@ -1,8 +1,12 @@
 import express from "express";
+import multer from "multer";
 import dataController from "../controllers/data.controller";
 import AccountsHelper from "../helpers/AccountsHelper";
 
+import config from "../../config/env";
+
 const router = express.Router(); // eslint-disable-line new-cap
+const upload = multer({ dest: config.express.uploadsTempLocation || "tmp/" });
 const keycloak = AccountsHelper.keycloakInstance();
 /**
  * @swagger
@@ -51,29 +55,26 @@ router.route("/create").post(dataController.create);
 
 /**
  * @swagger
- * /data/demo/{folder}:
+ * /data/bulk:
  *   post:
  *     tags:
  *       - Data
  *     description: Load Demo Data
+ *     consumes:
+ *       - multipart/form-data
  *     produces:
  *       - application/json
  *     parameters:
- *       - in: path
- *         name: folder
- *         required: true
+ *       - in: formData
+ *         name: file
+ *         description: The file to upload
  *         type: string
- *         description: The folder name to load from
- *       - in: body
+ *         format: binary
+ *         required: true
+ *       - in: formData
  *         name: purge
  *         description: A flag to purge existing experiments
- *         schema:
- *           type: object
- *           properties:
- *             purge:
- *               type: boolean
- *           example:
- *             purge: true
+ *         type: boolean
  *     security:
  *       - Bearer: []
  *     responses:
@@ -82,6 +83,6 @@ router.route("/create").post(dataController.create);
  *         schema:
  *           $ref: '#/definitions/BasicResponse'
  */
-router.route("/demo/:folder").post(keycloak.connect.protect(), dataController.loadDemo);
+router.route("/bulk").post(keycloak.connect.protect(), upload.single("file"), dataController.bulk);
 
 export default router;
