@@ -1,7 +1,15 @@
 import express from "express";
+
+import { jsonschema, request } from "makeandship-api-common/lib/modules/express/middleware";
+import * as schemas from "mykrobe-atlas-jsonschema";
+
 import AccountsHelper from "../helpers/AccountsHelper";
+import OrganisationHelper from "../helpers/OrganisationHelper";
+
 import organisationController from "../controllers/organisation.controller";
+import userController from "../controllers/user.controller";
 import config from "../../config/env";
+import Constants from "../Constants";
 
 const router = express.Router(); // eslint-disable-line new-cap
 const keycloak = AccountsHelper.keycloakInstance();
@@ -18,15 +26,50 @@ const keycloak = AccountsHelper.keycloakInstance();
  *         properties:
  *           name:
  *             type: string
- *           template:
+ *           slug:
  *             type: string
+ *           groupId:
+ *             type: string
+ *           owners:
+ *             type: array
+ *           members:
+ *             type: array
+ *           unapprovedMembers:
+ *             type: array
+ *           rejectedMembers:
+ *             type: array
  *           id:
  *             type: string
  *     example:
  *       status: success
  *       data:
  *         name: Apex Entertainment
- *         template: Apex template
+ *         slug: apex-entertainment
+ *         groupId: a842a3f45-ae66-41c5-b41c-798bc5e47a5b
+ *         owners:
+ *           - id: 890624089182796462cb1322
+ *             firtsname: Sam
+ *             lastname: Smith
+ *             email: sam@apex.com
+ *             username: sam@apex.com
+ *         members:
+ *           - id: 890624089182796462cb1322
+ *             firtsname: Sam
+ *             lastname: Smith
+ *             email: sam@apex.com
+ *             username: sam@apex.com
+ *         unapprovedMembers:
+ *           - id: 457624089182796462cgr666
+ *             firtsname: John
+ *             lastname: Thomas
+ *             email: john@apex.com
+ *             username: john@apex.com
+ *         rejectedMembers:
+ *           - id: 457624089182796462cy7g66
+ *             firtsname: Ali
+ *             lastname: Wood
+ *             email: ali@apex.com
+ *             username: ali@apex.com
  *         id: 588624076182796462cb133e
  */
 /**
@@ -43,15 +86,48 @@ const keycloak = AccountsHelper.keycloakInstance();
  *           properties:
  *             name:
  *               type: string
- *             template:
+ *             slug:
  *               type: string
+ *             groupId:
+ *               type: string
+ *             owners:
+ *               type: array
+ *             members:
+ *               type: array
+ *             unapprovedMembers:
+ *               type: array
  *             id:
  *               type: string
  *     example:
  *       status: success
  *       data:
  *         - name: Apex Entertainment
- *           template: Apex template
+ *           slug: apex-entertainment
+ *           groupId: a842a3f45-ae66-41c5-b41c-798bc5e47a5b
+ *           owners:
+ *             - id: 890624089182796462cb1322
+ *               firtsname: Sam
+ *               lastname: Smith
+ *               email: sam@apex.com
+ *               username: sam@apex.com
+ *           members:
+ *             - id: 890624089182796462cb1322
+ *               firtsname: Sam
+ *               lastname: Smith
+ *               email: sam@apex.com
+ *               username: sam@apex.com
+ *           unapprovedMembers:
+ *             - id: 457624089182796462cgr666
+ *               firtsname: John
+ *               lastname: Thomas
+ *               email: john@apex.com
+ *               username: john@apex.com
+ *           rejectedMembers:
+ *             - id: 457624089182796462cy7g66
+ *               firtsname: Ali
+ *               lastname: Wood
+ *               email: ali@apex.com
+ *               username: ali@apex.com
  *           id: 588624076182796462cb133e
  */
 router
@@ -98,18 +174,20 @@ router
    *           properties:
    *             name:
    *               type: string
-   *             template:
-   *               type: string
    *           example:
-   *             firstname: Apex Entertainment
-   *             lastname: Apex template
+   *             name: Apex Entertainment
    *     responses:
    *       200:
    *         description: Organisation data
    *         schema:
    *           $ref: '#/definitions/OrganisationResponse'
    */
-  .post(keycloak.connect.protect(), organisationController.create);
+  .post(
+    keycloak.connect.protect(),
+    jsonschema.schemaValidation(schemas["organisation"]),
+    userController.loadCurrentUser,
+    organisationController.create
+  );
 
 router
   .route("/:id")
@@ -164,18 +242,19 @@ router
    *           properties:
    *             name:
    *               type: string
-   *             template:
-   *               type: string
    *           example:
    *             name: Apex Entertainment
-   *             template: Apex template
    *     responses:
    *       200:
    *         description: Organisation data
    *         schema:
    *           $ref: '#/definitions/OrganisationResponse'
    */
-  .put(keycloak.connect.protect(), organisationController.update)
+  .put(
+    keycloak.connect.protect(),
+    request.whitelist(Constants.ORGANISATION_WHITELIST_FIELDS),
+    organisationController.update
+  )
   /**
    * @swagger
    * /organisations/{id}:
@@ -201,6 +280,244 @@ router
    *           $ref: '#/definitions/BasicResponse'
    */
   .delete(keycloak.connect.protect(), organisationController.remove);
+
+router
+  .route("/:id/join")
+  /**
+   * @swagger
+   * /organisations/{id}/join:
+   *   post:
+   *     tags:
+   *       - Organisations
+   *     description: Join an organisation
+   *     operationId: joinOrganisation
+   *     produces:
+   *       - application/json
+   *     security:
+   *       - Bearer: []
+   *     parameters:
+   *       - in: path
+   *         name: id
+   *         required: true
+   *         type: string
+   *         description: The organisation id
+   *     responses:
+   *       200:
+   *         description: A jsend response
+   *         schema:
+   *           $ref: '#/definitions/OrganisationResponse'
+   */
+  .post(
+    keycloak.connect.protect(),
+    userController.loadCurrentUser,
+    OrganisationHelper.checkInLists(Constants.ERRORS.JOIN_ORGANISATION, [
+      "unapprovedMembers",
+      "rejectedMembers",
+      "members"
+    ]),
+    organisationController.join
+  );
+
+router
+  .route("/:id/members/:memberId/approve")
+  /**
+   * @swagger
+   * /organisations/{id}/members/{memberId}/approve:
+   *   post:
+   *     tags:
+   *       - Organisations
+   *     description: Approve a join organisation request
+   *     operationId: approveJoinOrganisationRequest
+   *     produces:
+   *       - application/json
+   *     security:
+   *       - Bearer: []
+   *     parameters:
+   *       - in: path
+   *         name: id
+   *         required: true
+   *         type: string
+   *         description: The organisation id
+   *       - in: path
+   *         name: memberId
+   *         required: true
+   *         type: string
+   *         description: The member id
+   *     responses:
+   *       200:
+   *         description: A jsend response
+   *         schema:
+   *           $ref: '#/definitions/OrganisationResponse'
+   */
+  .post(
+    keycloak.connect.protect(),
+    userController.loadCurrentUser,
+    OrganisationHelper.isOwner(Constants.ERRORS.APPROVE_MEMBER),
+    OrganisationHelper.checkInLists(Constants.ERRORS.APPROVE_MEMBER, ["members"]),
+    OrganisationHelper.checkNotInLists(Constants.ERRORS.APPROVE_MEMBER, [
+      "unapprovedMembers",
+      "rejectedMembers"
+    ]),
+    organisationController.approve
+  );
+
+router
+  .route("/:id/members/:memberId/reject")
+  /**
+   * @swagger
+   * /organisations/{id}/members/{memberId}/reject:
+   *   post:
+   *     tags:
+   *       - Organisations
+   *     description: Reject a join organisation request
+   *     operationId: rejectJoinOrganisationRequest
+   *     produces:
+   *       - application/json
+   *     security:
+   *       - Bearer: []
+   *     parameters:
+   *       - in: path
+   *         name: id
+   *         required: true
+   *         type: string
+   *         description: The organisation id
+   *       - in: path
+   *         name: memberId
+   *         required: true
+   *         type: string
+   *         description: The member id
+   *     responses:
+   *       200:
+   *         description: A jsend response
+   *         schema:
+   *           $ref: '#/definitions/OrganisationResponse'
+   */
+  .post(
+    keycloak.connect.protect(),
+    userController.loadCurrentUser,
+    OrganisationHelper.isOwner(Constants.ERRORS.REJECT_MEMBER),
+    OrganisationHelper.checkInLists(Constants.ERRORS.REJECT_MEMBER, ["members", "rejectedMembers"]),
+    OrganisationHelper.checkNotInLists(Constants.ERRORS.REJECT_MEMBER, ["unapprovedMembers"]),
+    organisationController.reject
+  );
+
+router
+  .route("/:id/members/:memberId/remove")
+  /**
+   * @swagger
+   * /organisations/{id}/members/{memberId}/remove:
+   *   post:
+   *     tags:
+   *       - Organisations
+   *     description: Remove an organisation member
+   *     operationId: removeOrganisationMember
+   *     produces:
+   *       - application/json
+   *     security:
+   *       - Bearer: []
+   *     parameters:
+   *       - in: path
+   *         name: id
+   *         required: true
+   *         type: string
+   *         description: The organisation id
+   *       - in: path
+   *         name: memberId
+   *         required: true
+   *         type: string
+   *         description: The member id
+   *     responses:
+   *       200:
+   *         description: A jsend response
+   *         schema:
+   *           $ref: '#/definitions/OrganisationResponse'
+   */
+  .post(
+    keycloak.connect.protect(),
+    userController.loadCurrentUser,
+    OrganisationHelper.isOwner(Constants.ERRORS.REMOVE_MEMBER),
+    OrganisationHelper.checkNotInLists(Constants.ERRORS.REMOVE_MEMBER, ["members"]),
+    organisationController.removeMember
+  );
+
+router
+  .route("/:id/members/:memberId/promote")
+  /**
+   * @swagger
+   * /organisations/{id}/members/{memberId}/promote:
+   *   post:
+   *     tags:
+   *       - Organisations
+   *     description: Promote an organisation member
+   *     operationId: promoteOrganisationMember
+   *     produces:
+   *       - application/json
+   *     security:
+   *       - Bearer: []
+   *     parameters:
+   *       - in: path
+   *         name: id
+   *         required: true
+   *         type: string
+   *         description: The organisation id
+   *       - in: path
+   *         name: memberId
+   *         required: true
+   *         type: string
+   *         description: The member id
+   *     responses:
+   *       200:
+   *         description: A jsend response
+   *         schema:
+   *           $ref: '#/definitions/OrganisationResponse'
+   */
+  .post(
+    keycloak.connect.protect(),
+    userController.loadCurrentUser,
+    OrganisationHelper.isOwner(Constants.ERRORS.PROMOTE_MEMBER),
+    OrganisationHelper.checkNotInLists(Constants.ERRORS.PROMOTE_MEMBER, ["members"]),
+    organisationController.promote
+  );
+
+router
+  .route("/:id/owners/:memberId/demote")
+  /**
+   * @swagger
+   * /organisations/{id}/owners/{memberId}/demote:
+   *   post:
+   *     tags:
+   *       - Organisations
+   *     description: Demote an organisation owner
+   *     operationId: demoteOrganisationOwner
+   *     produces:
+   *       - application/json
+   *     security:
+   *       - Bearer: []
+   *     parameters:
+   *       - in: path
+   *         name: id
+   *         required: true
+   *         type: string
+   *         description: The organisation id
+   *       - in: path
+   *         name: memberId
+   *         required: true
+   *         type: string
+   *         description: The owner id
+   *     responses:
+   *       200:
+   *         description: A jsend response
+   *         schema:
+   *           $ref: '#/definitions/OrganisationResponse'
+   */
+  .post(
+    keycloak.connect.protect(),
+    userController.loadCurrentUser,
+    OrganisationHelper.isOwner(Constants.ERRORS.DEMOTE_MEMBER),
+    OrganisationHelper.checkNotInLists(Constants.ERRORS.DEMOTE_MEMBER, ["owners"]),
+    OrganisationHelper.checkEmptyList(Constants.ERRORS.DEMOTE_MEMBER, "owners"),
+    organisationController.demote
+  );
 
 /** Load user when API with id route parameter is hit */
 router.param("id", organisationController.load);
