@@ -1,5 +1,7 @@
 import AnalysisEventJSONTransformer from "./AnalysisEventJSONTransformer";
 
+import Constants from "../../Constants";
+
 /**
  * A class to transform event payloads
  * @property response : the response Object from mongoose
@@ -10,7 +12,7 @@ class DistanceStartedJSONTransformer extends AnalysisEventJSONTransformer {
    */
   transform(o, options) {
     const res = super.transform(o, options);
-    res.event = "Distance search started";
+    res.event = Constants.EVENTS.DISTANCE_SEARCH_STARTED.NAME;
     return res;
   }
 }
