@@ -1,5 +1,7 @@
 import ThirdPartyUploadEventJSONTransformer from "./ThirdPartyUploadEventJSONTransformer";
 
+import Constants from "../../Constants";
+
 /**
  * A class to transform event payloads
  * @property response : the response Object from mongoose
@@ -10,7 +12,9 @@ class ThirdPartyUploadCompleteJSONTransformer extends ThirdPartyUploadEventJSONT
    */
   transform(o, options) {
     const res = super.transform(o, options);
-    res.event = "Upload via 3rd party complete";
+
+    res.event = Constants.EVENTS.THIRD_PARTY_UPLOAD_COMPLETE.NAME;
+
     return res;
   }
 }
