@@ -1,5 +1,7 @@
 import UploadEventJSONTransformer from "./UploadEventJSONTransformer";
 
+import Constants from "../../Constants";
+
 /**
  * A class to transform event payloads
  * @property response : the response Object from mongoose
@@ -10,7 +12,9 @@ class UploadCompleteJSONTransformer extends UploadEventJSONTransformer {
    */
   transform(o, options) {
     const res = super.transform(o, options);
-    res.event = "Upload complete";
+
+    res.event = Constants.EVENTS.UPLOAD_COMPLETE.NAME;
+
     return res;
   }
 }
