@@ -1,6 +1,7 @@
 import request from "supertest";
 import httpStatus from "http-status";
-import swaggerParser from "swagger-parser";
+
+import Constants from "../../src/server/Constants";
 
 import { createApp } from "../setup";
 
@@ -8,7 +9,13 @@ import User from "../../src/server/models/user.model";
 
 import users from "../fixtures/users";
 
-const app = createApp();
+const args = {
+  app: null
+};
+
+beforeAll(async () => {
+  args.app = await createApp();
+});
 
 beforeEach(async () => {
   const userData = new User(users.admin);
@@ -16,13 +23,13 @@ beforeEach(async () => {
 });
 
 afterEach(async () => {
-  await User.remove({});
+  await User.deleteMany({});
 });
 
 describe("Misc", () => {
   describe("# GET /health-check", () => {
     it("should return OK", done => {
-      request(app)
+      request(args.app)
         .get("/health-check")
         .expect(httpStatus.OK)
         .end((err, res) => {
@@ -35,11 +42,13 @@ describe("Misc", () => {
 
   describe("# GET /404", () => {
     it("should return 404 status", done => {
-      request(app)
+      request(args.app)
         .get("/404")
         .expect(httpStatus.NOT_FOUND)
         .end((err, res) => {
-          expect(res.body.message).toEqual("Unknown API route.");
+          expect(res.body.status).toEqual("error");
+          expect(res.body.code).toEqual(Constants.ERRORS.ROUTE_NOT_FOUND);
+          expect(res.body.message).toEqual("Unknown API route");
           done();
         });
     });
@@ -47,11 +56,11 @@ describe("Misc", () => {
 
   describe("# Error Handling", () => {
     it("should handle mongoose CastError - Cast to ObjectId failed", done => {
-      request(app)
+      request(args.app)
         .get("/users/56z787zzz67fc")
         .expect(httpStatus.INTERNAL_SERVER_ERROR)
         .end((err, res) => {
-          expect(res.body.code).toEqual(10001);
+          expect(res.body.code).toEqual(Constants.ERRORS.GET_USER);
           expect(res.body.message).toEqual(
             'Cast to ObjectId failed for value "56z787zzz67fc" at path "_id" for model "User"'
           );
@@ -60,7 +69,7 @@ describe("Misc", () => {
     });
 
     it("should handle express validation error - email is required", done => {
-      request(app)
+      request(args.app)
         .post("/users")
         .send({
           firstname: "Yassire",
@@ -71,8 +80,7 @@ describe("Misc", () => {
         .expect(httpStatus.BAD_REQUEST)
         .end((err, res) => {
           expect(res.body.status).toEqual("error");
-          expect(res.body.code).toEqual(10005);
-          expect(res.body.data.errors[""].message).toEqual(
+          expect(res.body.data.errors.username.message).toEqual(
             "should have required property 'username'"
           );
           done();
