@@ -1,10 +1,13 @@
+import Constants from "../../src/server/Constants";
+
 import Organisation from "../../src/server/models/organisation.model";
 
-import setup from "../setup";
+import { createApp } from "../setup";
 
 import organisations from "../fixtures/organisations";
 
 let id = null;
+createApp();
 
 beforeEach(async done => {
   const organisationData = new Organisation(organisations.apex);
@@ -13,8 +16,8 @@ beforeEach(async done => {
   done();
 });
 
-afterEach(done => {
-  Organisation.remove({}, done);
+afterEach(async () => {
+  await Organisation.deleteMany({});
 });
 
 describe("## Organisations Functions", () => {
@@ -24,6 +27,7 @@ describe("## Organisations Functions", () => {
         const organisationData = new Organisation(organisations.diagnostics);
         const savedOrganisation = await organisationData.save();
         expect(savedOrganisation.name).toEqual("Diagnostic systems");
+        expect(savedOrganisation.slug).toEqual("diagnostic-systems");
         done();
       });
     });
@@ -33,6 +37,7 @@ describe("## Organisations Functions", () => {
       it("should return the organisation", async done => {
         const foundOrganisation = await Organisation.get(id);
         expect(foundOrganisation.name).toEqual("Apex Entertainment");
+        expect(foundOrganisation.slug).toEqual("apex-entertainment");
         done();
       });
     });
@@ -42,7 +47,7 @@ describe("## Organisations Functions", () => {
           await Organisation.get("58d3f3795d34d121805fdc61");
           fail();
         } catch (e) {
-          expect(e.name).toEqual("ObjectNotFound");
+          expect(e.code).toEqual(Constants.ERRORS.GET_ORGANISATION);
           expect(e.message).toEqual("Organisation not found with id 58d3f3795d34d121805fdc61");
           done();
         }
@@ -54,6 +59,7 @@ describe("## Organisations Functions", () => {
       const foundOrganisation = await Organisation.get(id);
       const json = foundOrganisation.toJSON();
       expect(json.name).toEqual("Apex Entertainment");
+      expect(json.slug).toEqual("apex-entertainment");
       done();
     });
   });
@@ -63,12 +69,13 @@ describe("## Organisations Functions", () => {
         const organisations = await Organisation.list();
         expect(organisations.length).toEqual(1);
         expect(organisations[0].name).toEqual("Apex Entertainment");
+        expect(organisations[0].slug).toEqual("apex-entertainment");
         done();
       });
     });
     describe("when no organisations exist", () => {
-      beforeEach(done => {
-        Organisation.remove({}, done);
+      beforeEach(async () => {
+        await Organisation.deleteMany({});
       });
       it("should return an empty array", async done => {
         const organisations = await Organisation.list();
