@@ -26,8 +26,8 @@ beforeEach(async () => {
 });
 
 afterEach(async () => {
-  await Search.remove({});
-  await User.remove({});
+  await Search.deleteMany({});
+  await User.deleteMany({});
 });
 
 describe("Search", () => {
@@ -323,7 +323,7 @@ describe("Search", () => {
           await Search.get("58d3f3795d34d121805fdc61");
           fail();
         } catch (e) {
-          expect(e.name).toEqual("ObjectNotFound");
+          expect(e.code).toEqual(Constants.ERRORS.GET_SEARCH);
           expect(e.message).toEqual("Search not found with id 58d3f3795d34d121805fdc61");
           done();
         }
